refactor: 컨트롤러 테스트 리팩토링에 맞게 수정
